:recycle: Rename pages.js
